Move Data's TryInto<…> impls to heap/object.rs
